Cleanup and creating toggle for netplan

+++ script/cleanup.sh
@@ -56,27 +56,30 @@ dd if=/dev/zero of=/tmp/whitespace bs=1024 count=$count
 rm /tmp/whitespace
 
 # Whiteout /boot
-count=$(df --sync -kP /boot | tail -n1 | awk -F ' ' '{print $4}')
-let count--
-dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count
-rm /boot/whitespace
-
-echo '==> Clear out swap and disable until reboot'
-set +e
-swapuuid=$(/sbin/blkid -o value -l -s UUID -t TYPE=swap)
-case "$?" in
-    2|0) ;;
-    *) exit 1 ;;
-esac
-set -e
-if [ "x${swapuuid}" != "x" ]; then
-    # Whiteout the swap partition to reduce box size
-    # Swap is disabled till reboot
-    swappart=$(readlink -f /dev/disk/by-uuid/$swapuuid)
-    /sbin/swapoff "${swappart}"
-    dd if=/dev/zero of="${swappart}" bs=1M || echo "dd exit code $? is suppressed"
-    /sbin/mkswap -U "${swapuuid}" "${swappart}"
+if grep -q /boot /etc/fstab; then 
+  count=$(df --sync -kP /boot | tail -n1 | awk -F ' ' '{print $4}')
+  let count--
+  dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count
+  rm /boot/whitespace
 fi
+
+# I dont need no stinking swap in my templates
+# echo '==> Clear out swap and disable until reboot'
+# set +e
+# swapuuid=$(/sbin/blkid -o value -l -s UUID -t TYPE=swap)
+# case "$?" in
+#     2|0) ;;
+#     *) exit 1 ;;
+# esac
+# set -e
+# if [ "x${swapuuid}" != "x" ]; then
+#     # Whiteout the swap partition to reduce box size
+#     # Swap is disabled till reboot
+#     swappart=$(readlink -f /dev/disk/by-uuid/$swapuuid)
+#     /sbin/swapoff "${swappart}"
+#     dd if=/dev/zero of="${swappart}" bs=1M || echo "dd exit code $? is suppressed"
+#     /sbin/mkswap -U "${swapuuid}" "${swappart}"
+# fi
 
 # Zero out the free space to save space in the final image
 dd if=/dev/zero of=/EMPTY bs=1M  || echo "dd exit code $? is suppressed"
@@ -93,4 +96,4 @@ echo "==> Disk usage after cleanup"
 df -h
 
 echo "==> Remove Machine-id file"
-> /etc/machine-id
+true > /etc/machine-id
